V0.2.0 working with FeatureCollections instead of GeometryCollections
