+++ src/geojson-types.ts
@@ -1,8 +1,11 @@
 export type Coordinate = [number, number];
 
-export interface Polygon<TData extends Record<string, unknown>> {
-  type: "Polygon";
-  coordinates: Array<Array<Coordinate>>;
+export interface Feature<TData extends Record<string, unknown>> {
+  type: "Feature";
+  geometry: {
+    type: "Polygon"
+    coordinates: Array<Array<Coordinate>>;
+  },
   properties: {
     colourHexCode: string;
     data: TData | null;
@@ -10,6 +13,6 @@ export interface Polygon<TData extends Record<string, unknown>> {
 }
 
 export interface OutputGeoJSON<TData extends Record<string, unknown>> {
-  type: "GeometryCollection";
-  geometries: Array<Polygon<TData>>;
+  type: "FeatureCollection";
+  features: Array<Feature<TData>>;
 }
